src/pyproject.toml
[project]
name = "expense-tracker-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "blinker==1.9.0",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "cryptography==46.0.2",
    "deprecated==1.2.18",
    "deprecation==2.1.0",
    "dotenv==0.9.9",
    "fastapi>=0.119.0",
    "flask==3.1.0",
    "flask-bcrypt==1.0.1",
    "flask-jwt-extended==4.7.1",
    "flask-limiter==3.9.2",
    "flask-wtf==1.2.2",
    "gotrue==2.11.4",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "hyperframe==6.1.0",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jwcrypto==1.5.6",
    "limits==5.6.0",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "ordered-set==4.1.0",
    "packaging==25.0",
    "passlib>=1.7.4",
    "postgrest==0.19.3",
    "psycopg2-binary==2.9.11",
    "pycparser==2.23",
    "pydantic==2.12.0",
    "pydantic-core==2.41.1",
    "pygments==2.19.2",
    "pyjwt==2.3.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-jwt==4.1.0",
    "realtime==2.22.0",
    "requests==2.32.5",
    "rich==13.9.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "storage3==0.11.3",
    "strenum==0.4.15",
    "supabase==2.13.0",
    "supafunc==0.9.4",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "websockets==15.0.1",
    "werkzeug==3.1.3",
    "wrapt==1.17.3",
    "wtforms==3.2.1",
]
